backend/requirements.txt
flask
flask-cors
supabase
python-dotenv
pyjwt
openai
langchain
langchain-openai
requests
gunicorn
